pyproject.toml
[tool.poetry]
name = "spellstone-ingest"
version = "0.0.1"
description = "Spellstone"
authors = ["Denis Volk <the.denis.volk@gmail.com>"]
license = "MIT License"
readme = "README.md"
packages = []


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "^1.0"
prefect = "^2.20"
requests = "^2.28"
pydantic = "^2"
pyyaml = "^6.0"
types-pyyaml = "^6.0"
obsidiantools = "^0.10"
markdown = "^3.6"
python-frontmatter = "^1.1"
feedparser = "^6.0"
markdownify = "^0.13"
trafilatura = "^1.12"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3"
types-requests = "^2.28"
types-retry = "^0.9"
types-python-dateutil = "^2.8"
jupyterlab = "^4.0"
ipykernel = "^6.22"
matplotlib = "^3.7"
pandas = "^2.0"
seaborn = "^0.12"
ipympl = "^0.9"
plotly = "^5.14"
tqdm = "^4.65"
pylint = "^2.17"
pylint-pydantic = "^0.2"
types-decorator = "^5.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line_length = 119
target_version = ['py310']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.github
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | notebooks
    | \.ipynb_checkpoints
    | __pycache__
    | local
    | data
    | logs
    | _build
    | buck-out
    | build
    | dist
    | snapshots
  )/
)
'''

[tool.pytest.ini_options]
markers = [
    "integration: integration testing (tests require valid credentials and can be slow)",
]
